+++ plugins/firebase/package.json
@@ -52,6 +52,7 @@
     "firebase-admin": "^13.2.0",
     "firebase-functions": "^6.3.2",
     "firebase-functions-test": "^3.4.0",
+    "firebase-tools": "^13.32.0",
     "fkill": "^9.0.0",
     "hono": "^4.6.3",
     "mime": "^4.0.4",
