+++ plugins/firebase/vitest.config.mts
@@ -4,5 +4,6 @@ export default defineConfig({
   test: {
     fileParallelism: false,
     globals: true,
+    globalSetup: ["./vitest.global-setup.ts"],
   },
 });
